Employee
Define Employee class with Employee code, name, date of birth and date of appointment. The Employee code must have the format of year-designation-number. The year is a two-digit integer such as 87. the designation is a single letter code M for manager, A for Administrative staff, H for HR dept staff, E for Executive staff, and T for Technical staff. The number is a three-digit number. The following are some sample employee codes.
82-M-183
76-A-242
71-H-107
Write a Java program to read the employee code, name, date of birth, and date of appointment and validate the employee code. If the employee code is incorrect a suitable user defined exception must be thrown. Then verify if date of birth is before date of appointment. If it is not so, then throw another user defined Exception. If it is correct, then create the Employee object, display the count of employee and display the details of employees